Switch binary from Ruby to Node powered
Now that the import-js npm module is almost ready, we can switch to it.
The biggest difference from the old binary is that commands are invoked
through a positional argument. The way unresolved imports are handled
has also changed. Instead of replaying the command with selections made,
you call `importjs add` directly.

I've updated documentation to instruct people to install the node
module. I expect us to bump version to 1.0.0 when we release the npm
module, so I've used that version in the docs.