Only show no executable error for the importjs executable error
I was seeing our no executable error message because I was working in a
project that had a directory in it that no longer existed. This was
because the subprocess couldn't find the cwd, not the importjs
executable. This confused me a little, so I think that scoping the error
message to only be shown if the executable is in the error string will
help.

Partly addresses #2.