Grab PATH on plugin initialization
I've started using nodenv recently. When I did, the Sublime import-js
plugin stopped working. This is because the `node` executable is no
longer available in /usr/local/bin. To fix this, we need to grab the
PATH environment variable from the user's shell (which should contain
the folder where `node` is available. )

I've shamelessly copied the approach that the SublimeLinter 3 plugin
uses ([1]). I simplified it however to only support posix systems and
(probably) also only bash and zsh. The technique involves executing a
command using the user's shell `echo`ing the $PATH variable. We can then
use that when we later execute `importjs`. The PATH extraction is only
run on plugin initialization (`plugin_loaded` [2]).

Fixes #2

[1]: https://github.com/SublimeLinter/SublimeLinter3/blob/88578d6256652862f900bc342f51f74c161f646e/lint/util.py
[2]: https://www.sublimetext.com/docs/3/api_reference.html