Use shell=True for subprocess command on Windows
subprocess.popen treats the PATH env variable differently on posix
systems and Windows. In short, PATH on posix is used to find the
executable (in our case `importjs`). PATH on Windows is not used to find
the executable, but it's been made available to the executable. Python
has an open bug [1] on the issue.

I couldn't find a better way to fix this than to turn on shell=True.
It has security issues, so it should be avoided. But I couldn't find
another way, and I don't want to spend much more time digging.

[1]: http://bugs.python.org/issue8557