Use importjsd daemon for importing
The latest version of import-js will have a new `importjsd` process you
can run in the background. By sending commands (in json format) to its
stdin, you can expect responses on stdout. This is a great speed
improvement for import-js, since we no longer have to spin up a node
environment every time you import.

Because we need the root folder of the current project when you start
the daemon, we delay daemon start until you make your first import. This
will make the first call a little slower, but on the other hand we avoid
starting the daemon if it's not needed.

I played around a bit with trying to make the whole process
asynchronous. In the end, the synchronous model ended up being a lot
simpler, and importing is so fast that you won't really notice that
we're blocking the main thread.