Switch to `--parent-pid` when starting daemon
I'm standardizing this across editors. It will allow the daemon process
to figure out for itself whether the parent process is still alive.

I'm doing this because there doesn't seem to be a stable way of getting
an exit signal from Sublime when it closes. So I'm left with zombie
importjsd processes.