Start separate daemons for each project folder
I started testing the Sublime plugin a little bit more and realized that
we need to handle the case where multiple projects are open at the same
time. I first started making the importjsd process able to switch cwd at
any time, but that turned out to be quite involved. So I started taking
a simpler way out and just start separate daemons for each project.
That's what you're seeing in this commit.