Update unresolved imports flow for import-js 2.7.0
A recent commit in import-js made it so that we send down more
information about unresolved imports:
Galooshi/import-js@194ed52

This was in service of allowing a module to export named export equal to
the module itself. Before said commit, we didn't have enough information
in the json object sent down to the client to target the right import.
Now, we have a full object describing the import.

Same as
https://github.com/Galooshi/atom-import-js/commit/2e47e808cc2d464442054841e9bbf8a83b162c8b
and
https://github.com/Galooshi/vim-import-js/commit/862ca76e685e8563385b0dcd8f7d83e47ba2ce09